+++ src/cmem/react-flow/index.ts
@@ -1,4 +1,3 @@
 export * from "./ReactFlow/ReactFlow";
-export * from "./ReactFlow/ReactFlowV10";
 export * from "./StickyNoteModal/StickyNoteModal";
 export * from "./extensions/scrollOnDragHook";