+++ src/extensions/react-flow/index.ts
@@ -14,6 +14,7 @@ export * from "./markers/ReactFlowMarkers";
 export * from "./minimap/MiniMap";
 export * from "./minimap/MiniMapV10";
 export * from "./minimap/utils";
+export * from "./versionsupport";
 
 // deprecated exports
 export { nodeTypes } from "./nodes/nodeTypes"; // @deprecated removed in v25